fix invalid one time login links
Code wasn't loading the Drupal user object but rather just casting the user registration array to an object, civicrm_entity wasn't picking up on the fact that the user object had hashed the supplied password. The one time login link was being created with the hashed password by Drupal, but then was clobbering the hashed password with the plaintext one later on and thus invalidating the one time login link.