Fix version constraints.
This fixes the version constraints for `civicrm/civicrm-drupal-8` as this is now a versioned repository (branches back to 5.0 exist).

Currently installs not using the `dev-master` of that repository will fail but using that constraint is dangerous as `dev-master` could contain (and has in the past!) bleeding edge code not compatible with the current version of `civicrm/civicrm-core`. 

This change pins both `civicrm/civicrm-core` and `civicrm/civicrm-drupal-8` to `~5.19` (any 5.x version of 5.19 or higher).

This also removes the repository entries for `civicrm-core` and `civicrm-drupal-8` these are ignored in any composer.json file other than the "root" (or "project") composer.json file anyway!
Fix trailing comma!
Fix trailing }