Update Hook signatures when invoking hooks
The signature used when calling CRM_Utils_Hook->invoke is deprecated, and from CiviCRM 5.29 this gives rise to a log warning message every time it occurs.  This change updates the signature to the preferred version.  It is applied to CiviCRM Versions >= 5.28 because I do not know the backward compatibility situation.